move customer tests to their own file, finish lithium tests
