+++ src/test/platform/formulas/assets/events/raw-case-webhook.json
@@ -1,0 +1,17 @@
+{
+  "Status": "New",
+  "LastModifiedDate": "2017-08-30T18:58:31.000+0000",
+  "IsDeleted": false,
+  "Priority": "Medium",
+  "LastViewedDate": "2017-08-30T18:58:31.000+0000",
+  "LastReferencedDate": "2017-08-30T18:58:31.000+0000",
+  "IsClosed": false,
+  "SystemModstamp": "2017-08-30T18:58:31.000+0000",
+  "OwnerId": "005i0000001VNdeAAG",
+  "CreatedById": "005i0000001VNdeAAG",
+  "IsEscalated": false,
+  "CreatedDate": "2017-08-30T18:58:31.000+0000",
+  "CaseNumber": "00008597",
+  "Id": "5000H00000zizkuQAA",
+  "LastModifiedById": "005i0000001VNdeAAG"
+}
