+++ src/test/platform/formulas/assets/formulas/customer-formulas/lithium-sfdc-bulk1-instance.json
@@ -1,0 +1,11 @@
+{
+    "name": "Lithium Test - Sfdc bulk1",
+    "configuration": {
+      "canrun": "yes",
+      "email": "churros@churros.com",
+      "lastruntime": "none",
+      "endtime": "blah",
+      "formulainstanceid": "{REPLACE}",
+      "sfdc.instance.id": "{REPLACE}"
+    }
+}
