+++ src/test/platform/formulas/assets/formulas/customer-formulas/lithium-sfdc-bulk2-instance.json
@@ -1,0 +1,9 @@
+{
+    "name": "Lithium Test - Sfdc bulk2",
+    "configuration": {
+      "counter": "0",
+      "email": "churros@churros.com",
+      "lithiumlsw.instance.id": "{REPLACE}",
+      "sfdc.instance.id": "{REPLACE}"
+    }
+}
